Start a simplified version of the ‘jif’ package: ‘jiffy’
We’re rebuilding from the ground up in ‘jiffy’ to ensure there
is negligible shear bias when using simulated data in Roaster.
The idea is to start with the minimal code that works, build unit
tests, and then expand the functionality as needed. There is a
lot of cruft in jif now from multiple projects.